Apply tool result names across Fireworks models
