+++ front/lib/model_constructors/providers/fireworks/models/kimi_k3.test.ts
@@ -59,73 +59,6 @@ describe("Kimi K3 model configuration", () => {
     });
   });
 
-  it("includes tool names when replaying parallel tool results", () => {
-    const endpoint = new MoonshotAiKimiK3GlobalFireworksStream({
-      FIREWORKS_API_KEY: "test",
-    });
-    const payload = endpoint.buildRequestPayload(
-      {
-        conversation: {
-          system: [],
-          messages: [
-            {
-              role: "assistant",
-              type: "tool_call_request",
-              content: {
-                callId: "call_1",
-                toolName: "first_tool",
-                arguments: "{}",
-              },
-            },
-            {
-              role: "assistant",
-              type: "tool_call_request",
-              content: {
-                callId: "call_2",
-                toolName: "second_tool",
-                arguments: "{}",
-              },
-            },
-            {
-              role: "user",
-              type: "tool_call_result",
-              content: {
-                callId: "call_1",
-                toolName: "first_tool",
-                parts: [{ type: "text", text: "first result" }],
-                isError: false,
-              },
-            },
-            {
-              role: "user",
-              type: "tool_call_result",
-              content: {
-                callId: "call_2",
-                toolName: "second_tool",
-                parts: [{ type: "text", text: "second result" }],
-                isError: false,
-              },
-            },
-          ],
-        },
-      },
-      MoonshotAiKimiK3GlobalFireworksStream.configSchema.parse({})
-    );
-
-    expect(payload.messages).toContainEqual({
-      role: "tool",
-      name: "first_tool",
-      tool_call_id: "call_1",
-      content: "first result",
-    });
-    expect(payload.messages).toContainEqual({
-      role: "tool",
-      name: "second_tool",
-      tool_call_id: "call_2",
-      content: "second result",
-    });
-  });
-
   it("exposes no `none` tier and reaches Fireworks natively at `light`", () => {
     // K3 always thinks, so `none` is not a reachable tier. `light` must map to
     // Fireworks' `low` rather than dropping reasoning_effort, which is what
